Update jamf2snipe
Re-worked the section where an asset is created for the first time, changed around the login on how it creats an asset tag if one does not exist and how it knows where to pull it from the config file.

Also added a few more info and debug outputs for troubleshooting and/or the verbose output in that section.