Fixed some mobile device asset tag issues. (#34)
* Update jamf2snipe

* Update jamf2snipe

Clarified the debug message after the time is checked or if it is a new record.

* Update jamf2snipe

Re-worked the section where an asset is created for the first time, changed around the login on how it creats an asset tag if one does not exist and how it knows where to pull it from the config file.

Also added a few more info and debug outputs for troubleshooting and/or the verbose output in that section.

* Update jamf2snipe

Added asset tag update if the device is a mobile device (added new sub routine for this).

* Update jamf2snipe

Fixed minor syntax issue at the end.

* Update jamf2snipe

Fixed the api call in the update jamf asset sub routine. changed "mobiledevice" to "mobile_device" as it should be.

* Update jamf2snipe

Fixed an issue where the mobile device asset tag update was calling back to the computer asset tag update when it was rate limited.

* Update jamf2snipe

Re-added the computer check before creating the payload in line 609-614 to account for the difference in location of the model_identifiers.

* Update jamf2snipe

Made the check more explicit rather than a catch all.