Make dryrun more useful for debugging
This changes the `--dryrun` flag so that it runs all code, but doesn't update anything in jamf or snipeit.
It also introduces a `--connection_test` flag, which takes over the functionality of the old dryrun flag.

The main idea behind this is, that for debugging and development purposes it might be useful to run almost all code,
but stop short of updating the environment.